Bitpacked representation of GregorianNormalizedDate
Under the assumption that passing around large datetime representations
will be harder on the CPU than packing / unpacking components as parts
of a bit field for arithmetic operations, we store each component of the
date in a bit field of minimal size. That way the entire iso8601
DateTime can be represented as 64 bits + 16 for the nanoseconds.